requirements.txt
# X (Twitter) メディアダウンローダー (API キー不要)
gallery-dl==1.27.4

# 動画ダウンローダー (YouTube / TikTok / NicoNico 等)
yt-dlp>=2024.1.0

# 設定管理 (pydantic v2 ベース)
pydantic-settings==2.3.4
pydantic==2.8.2

# Web サーバー (ギャラリー UI)
flask>=3.0.0
python-dotenv>=1.0.0
# サムネイル生成
Pillow>=10.0.0

# テスト
pytest>=8.0.0
pytest-mock>=3.14.0
